Fixed Transfer Pokemon Fault; FortId Issue
FortID went from int to string. I figured it would be null if it wasn't
deployed.

I should have tested or referred to the protobuf spec for more info.
Either way, it appears to be correctly transfering again. Not sure if
this will cause issue if Pokemon are deployed at gyms.