Pokemon Moves
I am not sure if this is the right place to put this / type it.
But these will be helpful later when MAYBE  a logic for trasnfer after it check what moves the pokemon has.. 
for each pokemon i took the best combo and grabbed the id's of these moves from the protos. 
again, no idea if it's pushed in the right spot/written the write way. whats important there is the data that does not exist currently in necro.