Improved bottleneck when run out of pokeballs (#1280)
Due to the sleep in here this function can delay the walk to a pokestop up to (pokemons around you) * 3 sec. Not sure if it's worth to implement the calls. A check for EventNoPokeballs after the EncounterPokemon with a loop break might be better even though it's after the sleep.