Fix coordinate related things (i.e. sniping) for non-us users
Since many sites like PokeVision rely on coordinates and other decimals
being seperated with dots and not with comma. For example, coordinates
have to be 41.242 and not 41,242.
Many non-us locales do use a comma for those separations. This commit
sets the locale for the program fixed to en-US which solves the problem.

Signed-off-by: Lyrex <admin@lyrex.net>