GetBestBall tweaks
Added StrictPokeballUse, currently even if a pokemon does not match the
specified criteria(cp/iv), a ball will be used whenever available, this
allows you to turn that off so the pokemon is skipped if no matching
ball is available.

Added PokemonToUsePokeball list, this will only return pokeballs when
the pokemon is in this list, if StrictPokeballUse is false, will return
to default behavior.

Added ThrowLuckyBallFirst, this will always throw a pokeball first when
available, if the pokemon matches ultraball criteria, then a greatball
is thrown when available.