[FEATURE] Added a low level killswitch
This killswitch is simply to stop the bot when any major issues occur.

This may help users in the future understand when a fatal issue of the
bot has occurred. For example, if this system was implemented when
Unknown6 was pushed, we could have killed the bot and given a reason